+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/util/RetryPredicates.scala
@@ -35,6 +35,10 @@ object RetryPredicates {
     case _                       => false
   }
 
+  def gkeRetryPredicate: Throwable => Boolean = {
+    case e: io.grpc.StatusRuntimeException => e.getStatus.getCode == io.grpc.Status.Code.INTERNAL
+  }
+
   def combine(predicates: Seq[Throwable => Boolean]): Throwable => Boolean = { throwable =>
     predicates.map(_(throwable)).foldLeft(false)(_ || _)
   }